Swagger (#5)
* Swagger config.

* Update.

* Remove IntelliJ project files.

+++ app/src/main/java/hu/bme/r0uj46/cookbook/ui/details/DetailsPresenter.java
@@ -22,7 +22,7 @@ public class DetailsPresenter extends Presenter<DetailsScreen> {
         super.detach();
     }
 
-    public void loadRecipe(int recipeId) {
+    public void loadRecipe(Long recipeId) {
         // TODO
         screen.displayRecipe(recipesInteractor.getRecipe(recipeId));
     }
