+++ app/src/main/java/hu/bme/r0uj46/cookbook/ui/main/MainScreen.java
@@ -7,6 +7,6 @@ import hu.bme.r0uj46.cookbook.model.Recipe;
 public interface MainScreen {
     void listRecipes(List<Recipe> recipes);
 
-    void showRecipeDetails(int recipeId);
+    void showRecipeDetails(Long recipeId);
     void showNewRecipe();
 }
